ajout calcul dynamique du prix de l'inscription côté front (manque récupération des donnés BDD event)

+++ src/main/java/com/wildcodeschool/festivalorleansjoue/utils/MathUtils.java
@@ -2,12 +2,12 @@ package com.wildcodeschool.festivalorleansjoue.utils;
 
 public class MathUtils {
 	
-	public static float priceRound(float price) {		
+	public static float priceRound(double price) {		
 		return (float)(Math.round(price * 100)) / 100;
 	}
 	
 	public static float registrationCost(double tablesQuantity, double tablePrice, double saleOptionPrice ) {
-		float registrationCost = (float)(tablesQuantity * tablePrice - (double)(tablesQuantity / 4) * tablePrice + saleOptionPrice);
+		float registrationCost = (float)(tablesQuantity * tablePrice - (float)((int)(tablesQuantity / 4)) * tablePrice + saleOptionPrice);
 		return (float)(Math.round(registrationCost * 100)) / 100;
 	}
